requirements.txt
aiodns==2.0.0
aiohttp==3.6.3
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
autopep8==1.5.4
cchardet==2.1.7
cffi==1.14.4
chardet==3.0.4
discord.py==1.5.1
gino==1.0.1
hiredis==1.1.0
idna==2.10
multidict==4.7.6
pycares==3.1.1
pycodestyle==2.6.0
pycparser==2.20
PyYAML==5.3.1
rope==0.18.0
six==1.15.0
SQLAlchemy==1.3.15
SQLAlchemy-Utils==0.36.8
toml==0.10.2
yarl==1.5.1
